Speech_to_text/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic-settings>=2.0.0
redis>=4.5.0
aioredis>=2.0.0
tenacity>=9.1.2
celery>=5.3.0

torch>=2.2.0
torchaudio>=2.2.0
jiwer>=4.0.0
tensorboard==2.16.2
sentencepiece==0.1.99
tqdm==4.66.2
numpy==1.26.4
python-multipart==0.0.19
pyannote.audio==3.3.2